requirements.txt
yfinance
pandas
numpy
streamlit
plotly
pytz
requests
python-dotenv
